test: Extract common classes
Get rids of imports from other testfiles
