+++ unit_tests/test_gemini_thread.py
@@ -23,7 +23,7 @@ from sdcm.gemini_thread import GeminiStressThread
 from sdcm.utils.docker_utils import running_in_docker
 from unit_tests.conftest import configure_scylla_node
 from unit_tests.dummy_remote import LocalLoaderSetDummy
-from unit_tests.test_cluster import DummyDbCluster
+from unit_tests.lib.fake_cluster import DummyDbCluster
 
 pytestmark = [
     pytest.mark.usefixtures("events"),
