+++ unit_tests/test_python_driver.py
@@ -2,7 +2,7 @@ import logging
 
 import pytest
 
-from unit_tests.test_cluster import DummyDbCluster
+from unit_tests.lib.fake_cluster import DummyDbCluster
 
 
 log = logging.getLogger(__name__)
